Make projections trait based instead of class based
